% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeMetlinFun.R
\name{getMetlinMS2}
\alias{getMetlinMS2}
\title{Obtain Metlin MS/MS data for a given compound ID.}
\usage{
getMetlinMS2(cmpd_id)
}
\arguments{
\item{compound_id}{The Metlin database number corresponding to the compound
for which MS/MS data is desired}
}
\description{
\code{getMetlinMS2} queries Metlin for MS/MS data corresponding to a given
Metlin compound ID. This function returns a long-form data frame with the
containing all fragment masses and fragment intensities for a given polarity,
adduct, and voltage, as found in the database. The base url here is
https://metlin.scripps.edu/showChart.php?molid=[CMPD ID]&etype=experimental,
although the actual data is scraped from the Javascript behind the
interactive graph. If no MS2 data is found, getMetlinMS2 will tell you and
you should check this with the website.
}
\examples{
getMetlinMS2(35) #L-Valine
getMetlinMS2(287) #Betaine

}
