% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeMetlinFun.R
\name{getMetlinMz}
\alias{getMetlinMz}
\title{Obtain Metlin data for a given mz and ppm.}
\usage{
getMetlinMz(cmpd_mz, ppm = 2.5)
}
\arguments{
\item{cmpd_mz}{The mass of the compound you're searching for in Metlin. This
mass will be automatically expanded in to a window based on the provided
\code{ppm}.}

\item{ppm}{The ppm accuracy of the instrument that you've collected MS data
from. This will be used to expand the \code{cmpd_mz} provided above.}
}
\value{
A data frame containing Metlin compound ID, exact mass, compound
  name, formula, CAS and KEGG numbers, and the MSMS status of each compound
  (experimental vs. none). It does not currently return the structures.
}
\description{
\code{getMetlinMz} queries Metlin for a given mass and ppm. The default ppm
is 2.5. This function returns a data frame containing Metlin compound ID,
exact mass, compound name, formula, CAS and KEGG numbers, and the MSMS status
of each compound (experimental vs. none). It does not currently return
structures. This information is scraped from the base url
https://metlin.scripps.edu/advanced_search_result.php?molid=&mass_min=[MIN
MZ]&mass_max=[MAX MZ]&AminoAcid=add&drug=add&toxinEPA=add&keggIDFilter=add.
}
\examples{
getMetlinMz(117.078979) #Betaine
getMetlinMz(135.054495) #Adenine
getMetlinMz(mz = 135.054495, ppm = 500)

}
