% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeMetlinFun.R
\name{getMetlinName}
\alias{getMetlinName}
\title{Obtain Metlin data for a given compound name.}
\usage{
getMetlinName(name)
}
\arguments{
\item{name}{The name of the compound}
}
\value{
A data frame containing Metlin compound ID, exact mass, compound
  name, formula, CAS and KEGG numbers, and the MSMS status of each compound
  (experimental vs. none). It does not currently return the structures.
}
\description{
\code{getMetlinName} queries Metlin for a given compound name. This function
returns a data frame containing Metlin compound ID, exact mass, compound
name, formula, CAS and KEGG numbers, and the MSMS status of each compound
(experimental vs. none). It does not currently return structures. This
information is scraped from the base url
https://metlin.scripps.edu/advanced_search_result.php?name=[COMPOUND
NAME]&AminoAcid=add&drug=add&toxinEPA=add&keggIDFilter=add. This function
behaves identically to searching Metlin via the "Advanced Search" option and
providing only a name. It's a bit oversealous in its searching, and will
often pull up compounds that aren't even close. This is due to Metlin's
search functionality, not this package's internal workings.
}
\examples{
getMetlinMz("betaine")
getMetlinMz("adenine")

}
